requirements.txt
psycopg2-binary==2.9.10
requests
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2